refactor: rename pages folder and file
