+++ apps/freelance/src/pages/RecoverPasswordReset/index.tsx
@@ -1,0 +1,5 @@
+import { RecoverPasswordSecondForm } from "@freelance/components";
+
+export function RecoverPasswordReset() {
+	return <RecoverPasswordSecondForm />;
+}
