+++ apps/freelance/src/pages/RecoverPasswordReset/password-updated.jsx
@@ -1,0 +1,7 @@
+import { useTranslation } from "react-i18next";
+
+export const PasswordUpdated = () => {
+	const { t } = useTranslation();
+
+	return <h1>{t("recoverPassForm.passwordHasUpdated")}</h1>;
+};
