--- apps/freelance/src/pages/recoverpass-first-page/index.ts
@@ -1,2 +1,0 @@
-export { RecoverPassFirstPage } from './recoverpass-first-page';
-export { CheckYourEmail } from './check-your-email';
