--- apps/freelance/src/pages/recoverpass-third-page/password-updated.jsx
@@ -1,7 +1,0 @@
-import { useTranslation } from "react-i18next";
-
-export const PasswordUpdated = () => {
-  const {t} = useTranslation();
-  
-  return <h1>{t('recoverPassForm.passwordHasUpdated')}</h1>;
-};
