--- apps/freelance/src/pages/recoverpass-third-page/recoverpassword-third-page.tsx
@@ -1,5 +1,0 @@
-import { RecoverPasswordSecondForm } from '@freelance/components';
-
-export function RecoverPasswordThirdPage() {
-  return <RecoverPasswordSecondForm />;
-}
